Move wrappers to /api/types.
